First commited as before_46469_rev2

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/file/ObjectDirectoryPackParser.java
@@ -50,7 +50,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.RandomAccessFile;
-import java.nio.file.StandardCopyOption;
 import java.security.MessageDigest;
 import java.text.MessageFormat;
 import java.util.Arrays;
@@ -478,25 +477,20 @@ public class ObjectDirectoryPackParser extends PackParser {
 			}
 		}
 
-		try {
-			FileUtils.rename(tmpPack, finalPack,
-					StandardCopyOption.ATOMIC_MOVE);
-		} catch (IOException e) {
+		if (!tmpPack.renameTo(finalPack)) {
 			cleanupTemporaryFiles();
 			keep.unlock();
 			throw new IOException(MessageFormat.format(
-					JGitText.get().cannotMovePackTo, finalPack), e);
+					JGitText.get().cannotMovePackTo, finalPack));
 		}
 
-		try {
-			FileUtils.rename(tmpIdx, finalIdx, StandardCopyOption.ATOMIC_MOVE);
-		} catch (IOException e) {
+		if (!tmpIdx.renameTo(finalIdx)) {
 			cleanupTemporaryFiles();
 			keep.unlock();
 			if (!finalPack.delete())
 				finalPack.deleteOnExit();
 			throw new IOException(MessageFormat.format(
-					JGitText.get().cannotMoveIndexTo, finalIdx), e);
+					JGitText.get().cannotMoveIndexTo, finalIdx));
 		}
 
 		try {
